refactor(frontend/kafka-proxy): extract `EventName` to `@talus/model`
