+++ libs/model/src/index.ts
@@ -1,2 +1,3 @@
+export * from './lib/event-name.value';
 export * from './lib/rgba.value';
 export * from './lib/tool.value';
