+++ libs/model/src/lib/event-name.value.ts
@@ -1,0 +1,7 @@
+export enum EventName {
+  SyncAction = 'SyncAction',
+  ConsumeTopic = 'ConsumeTopic',
+  CreateTopic = 'CreateTopic',
+  DeleteTopic = 'DeleteTopic',
+  TopicNames = 'TopicNames',
+}
