+++ nx.json
@@ -36,7 +36,7 @@
       "tags": ["scope:libs"]
     },
     "kafka-proxy": {
-      "tags": []
+      "tags": ["scope:backend"]
     }
   }
 }
